fix(gms): raise inventory slot limit to 192 for ExpandItem
Cash and tab slot caps align with the client ExpandItem UI (96->192).

+++ gms-server/src/main/java/org/gms/net/server/channel/handlers/CashOperationHandler.java
@@ -198,7 +198,7 @@ public final class CashOperationHandler extends AbstractPacketHandler {
                         }
                         cs.gainCash(cash, -4000);
                         if (chr.gainSlots(type, qty, false)) {
-                            c.sendPacket(PacketCreator.showBoughtInventorySlots(type, chr.getSlots(type)));
+                            c.sendPacket(PacketCreator.showBoughtInventorySlots(type, (short) chr.getSlots(type)));
                             c.sendPacket(PacketCreator.showCash(chr));
                         } else {
                             log.warn("Could not add {} slots of type {} for chr {}", qty, type, Character.makeMapleReadable(chr.getName()));
@@ -221,7 +221,7 @@ public final class CashOperationHandler extends AbstractPacketHandler {
                         }
                         cs.gainCash(cash, cItem, chr.getWorld());
                         if (chr.gainSlots(type, qty, false)) {
-                            c.sendPacket(PacketCreator.showBoughtInventorySlots(type, chr.getSlots(type)));
+                            c.sendPacket(PacketCreator.showBoughtInventorySlots(type, (short) chr.getSlots(type)));
                             c.sendPacket(PacketCreator.showCash(chr));
                         } else {
                             log.warn("Could not add {} slots of type {} for chr {}", qty, type, Character.makeMapleReadable(chr.getName()));
